third_party/WebKit/LayoutTests/fast/canvas/canvas-arc-negative-radius-expected.txt

Tests CanvasPathMethods arc and arcTo with negative radii.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.arc(10, 10, 5, 0, 1, false) did not throw exception.
PASS ctx.arc(10, 10, 0, 0, 1, false) did not throw exception.
PASS ctx.arc(10, 10, -5, 0, 1, false) threw exception IndexSizeError: Failed to execute 'arc' on 'CanvasRenderingContext2D': The radius provided (-5) is negative..
PASS ctx.arcTo(10, 10, 20, 20, 5) did not throw exception.
PASS ctx.arcTo(10, 10, 20, 20, 0) did not throw exception.
PASS ctx.arcTo(10, 10, 20, 20, -5) threw exception IndexSizeError: Failed to execute 'arcTo' on 'CanvasRenderingContext2D': The radius provided (-5) is negative..
PASS path.arc(10, 10, 5, 0, 1, false) did not throw exception.
PASS path.arc(10, 10, 0, 0, 1, false) did not throw exception.
PASS path.arc(10, 10, -5, 0, 1, false) threw exception IndexSizeError: Failed to execute 'arc' on 'Path2D': The radius provided (-5) is negative..
PASS path.arcTo(10, 10, 20, 20, 5) did not throw exception.
PASS path.arcTo(10, 10, 20, 20, 0) did not throw exception.
PASS path.arcTo(10, 10, 20, 20, -5) threw exception IndexSizeError: Failed to execute 'arcTo' on 'Path2D': The radius provided (-5) is negative..
PASS successfullyParsed is true

TEST COMPLETE
